Moved method to hepers and added unit test for the helper method as per review suggestion
